Script that do a QGIS process through time
This script answers (somehow) the question from GIS_stackexchange: "Running and scheduling processing jobs in QGIS 3"                    *
Basically is a raw script that reads a layer (shp, gkpg) and performs a buffer iteratively over time. 
The interesting aproach is that the process is stoppable when time is waiting for next iteration.